+++ src/constants/auth.ts
@@ -1,0 +1,2 @@
+// 인증번호 입력 시간 상수화
+export const AUTH_TIMER_DURATION = 180;
